add HUD Canvas prefab
move DamageGauge into UI folder and namespace
