+++ Assets/Prefabs/HUD.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 14fc403ff02da1341b4242cb2261ff23
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
